[#7]/refactor: change directory name
